fixed all tests, broke portals test file into smaller test files
